refactor: 컨텍스트 사용하는 훅 분리
contexts/hooks에 보관
